chat-with-pdf/requirements.txt
streamlit
tiktoken
pymupdf
sentence-transformers
google-generativeai
python-dotenv